refactor: background.ts 제거
